// client/src/app/types/blog.ts
export interface BlogPost {
    id: string;
    title: string;
    slug: string;
    excerpt: string;
    content: string;
    image_url: string;
    date: string;
    read_time: string;
    category: string;
    author: string;
    author_bio: string;
    is_featured: boolean;
    created_at: string;
  }